% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subincomeR-package.R
\docType{package}
\name{subincomeR-package}
\alias{subincomeR}
\alias{subincomeR-package}
\title{subincomeR: Access to Global Sub-National Income Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides access to granular sub-national income data from the MCC-PIK Database Of Sub-national Economic Output (DOSE). The package downloads and processes the data from its open repository on 'Zenodo' (\url{https://zenodo.org/records/13773040}). Functions are provided to fetch data at multiple geographic levels, match coordinates to administrative regions, and access associated geometries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pablogguz/subincomeR}
  \item \url{https://pablogguz.github.io/subincomeR/}
  \item Report bugs at \url{https://github.com/pablogguz/subincomeR/issues}
}

}
\author{
\strong{Maintainer}: Pablo García Guzmán \email{garciagp@ebrd.com} [copyright holder]

}
\keyword{internal}
